Add feature to switch captcha validation on/off

+++ survey/apps/main/views.py
@@ -27,19 +27,20 @@ class UserViewSet(viewsets.ModelViewSet):
         """
         Create user if captcha is valid
         """
-        if not all(x in request.data for x in ['captcha_key', 'captcha_value']):
-            return Response(status=status.HTTP_400_BAD_REQUEST)
-        
-        captcha_key = request.data['captcha_key']
-        captcha_value = request.data['captcha_value']
-        
-        try:
-            captcha = CaptchaStore.objects.get(challenge=captcha_value, hashkey=captcha_key)
-            captcha.delete()
-        except:
-            return Response(
-                    {'state': False, 'code': 1, 'message': 'Captcha input is not correct.'},
-                    status=status.HTTP_200_OK)
+        if getattr(settings, 'USE_CAPTCHA') == True:
+            if not all(x in request.data for x in ['captcha_key', 'captcha_value']):
+                return Response(status=status.HTTP_400_BAD_REQUEST)
+            
+            captcha_key = request.data['captcha_key']
+            captcha_value = request.data['captcha_value']
+            
+            try:
+                captcha = CaptchaStore.objects.get(challenge=captcha_value, hashkey=captcha_key)
+                captcha.delete()
+            except:
+                return Response(
+                        {'state': False, 'code': 1, 'message': 'Captcha input is not correct.'},
+                        status=status.HTTP_200_OK)
         
         serializer = self.get_serializer(data=request.data)
         serializer.is_valid(raise_exception=True)
